refactor: move JSONSchema to jsonSchemaTypes
