+++ packages/schema/src/schema.ts
@@ -1,4 +1,3 @@
-import type { JSONSchema7 } from 'json-schema';
 import { isOptionalNullable, objectKeyEncode } from './utils';
 import type { JSONSchemaContext, PartialJSONSchema } from './jsonSchemaTypes';
 
@@ -26,21 +25,6 @@ export interface Schema<T, S = any> {
   unmapXml: (value: T, ctxt: SchemaContextCreator) => any;
 
   toJSONSchema: (context: JSONSchemaContext) => PartialJSONSchema;
-}
-
-/**
- * The equivalent JSON Schema representation of the Schema interface.
- * It targets the superset of JSON Schema 2020-12 specified in OpenAPI 3.1.0.
- * See: https://spec.openapis.org/oas/v3.1.0.html#data-types.
- * NOTE: Future compatibility with JSON Schema draft-07 is not guaranteed.
- * draft-07 was chosen as a base type since many existing libraries already use it.
- */
-export interface JSONSchema extends JSONSchema7 {
-  $schema: 'https://spec.openapis.org/oas/3.1/dialect/base';
-  discriminator?: {
-    propertyName: string;
-    mapping?: { [discriminatorValue: string]: string };
-  };
 }
 
 /**
